replicate: use bearer token instead of did
